Split builder, build_env and binary patcher into new builder module
