+++ src/builder/mod.rs
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: GPL-3.0-only
+mod build_env;
+mod builder;
+pub mod error;
+mod patcher;
+
+pub use self::builder::Builder;
+
+pub use self::build_env::BuildEnv;
+
+pub use self::patcher::BinaryPatcher;
+pub use self::patcher::BinaryPatcherError;
