+++ src/installer/error.rs
@@ -2,9 +2,9 @@
 use thiserror::Error;
 
 use crate::{
+    builder::error::BuilderError,
     cli::display::error::DisplayError,
     installer::{
-        builder::BuilderError,
         scripts::ScriptError,
         types::{PackageId, PackageIdError, Version},
         unpack::UnpackError,
